Move TestEnv to separate module.
